refactor(saveslot): move the slot system off the player into HasSaveSlotsDataStore
The save slot model -- the roster, per-slot stores, selection, create/delete/reset,
export/import, ephemeral slots and playtime -- now lives in HasSaveSlotsDataStore, built
over an injected datastore rather than a bound player. HasSaveSlots keeps only what needs a
player in this server (replicating the slot container, remotes, teleport data, summary
providers) and delegates the rest.

Nothing changes for callers: every method keeps its name, signature and behavior. The point
is that a slot is now reachable for a player who is not in this server, which is what the
offline admin tooling needs.

The four replicated value objects and the shared-store service are injected, the pre-select
veto and the share-code identity are hooks, and the slot container's parent is a value object
the owner mounts -- so the offline path mounts nothing and replicates to no one.
