+++ src/saveslot/src/Server/Binders/HasSaveSlots.lua
@@ -1,18 +1,27 @@
 --!strict
 --[=[
+	Binds the save slot system to a live [Player].
+
+	The slot model itself -- the roster, per-slot stores, selection, create/delete/reset,
+	export/import, ephemeral slots and playtime -- lives in [HasSaveSlotsDataStore], built here over
+	the player's datastore. What stays behind is everything that only means something for a player in
+	this server: replicating the slot container under them, the client remotes, teleport data, and
+	summary providers (which are handed the [Player]).
+
+	Every slot method below is a straight delegation. Reach for [HasSaveSlotsDataStore] directly only
+	when there is no player to bind -- admin tooling acting on someone who is not in this server.
+
 	@class HasSaveSlots
 ]=]
 
 local require = require(script.Parent.loader).load(script)
 
-local HttpService = game:GetService("HttpService")
-
 local Binder = require("Binder")
 local Brio = require("Brio")
 local DataStoreStage = require("DataStoreStage")
 local HasSaveSlotsBase = require("HasSaveSlotsBase")
+local HasSaveSlotsDataStore = require("HasSaveSlotsDataStore")
 local HasSaveSlotsInterface = require("HasSaveSlotsInterface")
-local InMemoryDataStore = require("InMemoryDataStore")
 local Maid = require("Maid")
 local Observable = require("Observable")
 local ObservableMap = require("ObservableMap")
@@ -21,71 +30,70 @@ local PlayerDataStoreService = require("PlayerDataStoreService")
 local Promise = require("Promise")
 local Remoting = require("Remoting")
 local Rx = require("Rx")
-local RxBrioUtils = require("RxBrioUtils")
 local SaveSlotCodeUtils = require("SaveSlotCodeUtils")
 local SaveSlotConstants = require("SaveSlotConstants")
 local SaveSlotData = require("SaveSlotData")
 local SaveSlotExportUtils = require("SaveSlotExportUtils")
 local SaveSlotSharedDataStoreService = require("SaveSlotSharedDataStoreService")
 local ServiceBag = require("ServiceBag")
+local Signal = require("Signal")
 local TeleportDataService = require("TeleportDataService")
+local ValueObject = require("ValueObject")
 
 -- A summary provider contributes one named piece of a slot's Summary. It is called with the player and
 -- the slot's DataStoreStage and returns an Observable of a JSON-serializable value. Providers are
 -- aggregated into a table keyed by their registered name; see RegisterSummaryProvider and _setupSummary.
 export type SummaryProvider = (Player, DataStoreStage.DataStoreStage) -> Observable.Observable<any>
 
+-- Re-exported so consumers keep naming this through HasSaveSlots; see HasSaveSlotsDataStore.
+export type SaveSlotCreateMetadata = HasSaveSlotsDataStore.SaveSlotCreateMetadata
+
 -- A provider that errors (when called, or mid-stream) contributes this sentinel rather than stalling or
 -- failing the whole aggregate; _setupSummary strips it back out before writing the Summary. Compared by
 -- identity, so it never collides with a real value (even an empty table) a provider might emit.
 local NONE = {}
 
--- The caller-supplied fields for a new slot. SlotId and SlotIndex are assigned by PromiseCreateSlot
--- itself (from a fresh GUID and the slotIndex argument), so they are never taken from here.
--- TimePlayed seeds the new slot's accrued playtime, which is what lets a copy of a slot (duplicate or
--- export/import) carry the playtime that belongs to the progress being copied.
-export type SaveSlotCreateMetadata = {
-	SlotName: string?,
-	Summary: SaveSlotData.SaveSlotSummary?,
-	TimePlayed: number?,
-}
-
-local HasSaveSlots = setmetatable({}, HasSaveSlotsBase)
+local HasSaveSlots = {}
 HasSaveSlots.ClassName = "HasSaveSlots"
 HasSaveSlots.__index = HasSaveSlots
+-- Runtime inheritance only. Typing the metatable target as `any` keeps the old solver from chasing
+-- the full HasSaveSlotsBase type here, which -- now that this class also names HasSaveSlotsDataStore
+-- across every delegation -- overflows its complexity budget ("Code is too complex"). The inherited
+-- surface is supplied structurally below.
+setmetatable(HasSaveSlots :: any, HasSaveSlotsBase)
+
+-- Minimal structural surface inherited from HasSaveSlotsBase, for this class and its consumers. See
+-- the setmetatable note above for why we don't intersect HasSaveSlotsBase.HasSaveSlotsBase.
+type HasSaveSlotsBaseLike = {
+	_obj: Player,
+	_maid: Maid.Maid,
+
+	ActiveSlotId: ValueObject.ValueObject<SaveSlotData.SlotId?>,
+	LastActiveSlotId: ValueObject.ValueObject<SaveSlotData.SlotId?>,
+	ActiveTransferableEphemeralKey: ValueObject.ValueObject<string?>,
+	MaxSlotCount: ValueObject.ValueObject<number>,
+	SlotChanged: Signal.Signal<SaveSlotData.SlotId>,
+
+	PromiseHasSaveSlotFromTeleport: (self: any) -> Promise.Promise<boolean>,
+	PromiseLoadSaveSlotFromTeleport: (self: any) -> Promise.Promise<SaveSlotData.SlotId?>,
+	Destroy: (self: any) -> (),
+}
 
 export type HasSaveSlots =
 	typeof(setmetatable(
 		{} :: {
-			_obj: Player,
 			_serviceBag: ServiceBag.ServiceBag,
 			_playerDataStoreService: any,
-			_slotContainer: Folder,
-			_slotMap: { [SaveSlotData.SlotId]: Folder },
-			-- The in-memory stores backing ephemeral (session-only) slots, keyed by slot id. A slot's
-			-- IsEphemeral property is the discriminator (see _isEphemeral); this holds the store objects a
-			-- property can't. See PromiseSelectEphemeralSlot.
-			_ephemeralStores: { [SaveSlotData.SlotId]: InMemoryDataStore.InMemoryDataStore },
-			-- The shared-store key each transferable ephemeral slot was loaded from, so a teleport can
-			-- re-save its live state under that key and carry it forward. See PromiseSelectTransferableEphemeralSlot.
-			_transferableEphemeralKeys: { [SaveSlotData.SlotId]: string },
-			_codeGenerator: SaveSlotCodeUtils.CodeGenerator,
-			_loadPromise: Promise.Promise<{}>,
+			-- A HasSaveSlotsDataStore. Untyped for the reason given on GetSlotsDataStore.
+			_slotsDataStore: any,
 			_remoting: any,
-			_dataStore: any,
-			_systemStore: any,
-			_metadataStore: any,
 			_summaryProviders: ObservableMap.ObservableMap<string, SummaryProvider>,
-			_lastActiveSlotId: SaveSlotData.SlotId?,
 			_teleportDataService: any,
 			_sharedSaveSlotDataStoreService: any,
-			_playSessionSlotId: SaveSlotData.SlotId?,
-			_playSessionStart: number?,
-			_playSessionLastFlush: number?,
 		},
 		{} :: typeof({ __index = HasSaveSlots })
 	))
-	& HasSaveSlotsBase.HasSaveSlotsBase
+	& HasSaveSlotsBaseLike
 
 function HasSaveSlots.new(player: Player, serviceBag: ServiceBag.ServiceBag): HasSaveSlots
 	local self: HasSaveSlots = setmetatable(HasSaveSlotsBase.new(player, serviceBag) :: any, HasSaveSlots)
@@ -95,24 +103,39 @@ function HasSaveSlots.new(player: Player, serviceBag: ServiceBag.ServiceBag): Ha
 	self._teleportDataService = self._serviceBag:GetService(TeleportDataService)
 	self._sharedSaveSlotDataStoreService = self._serviceBag:GetService(SaveSlotSharedDataStoreService)
 
-	self._slotContainer = self._maid:Add(Instance.new("Folder"))
-	self._slotContainer.Name = SaveSlotConstants.METADATA_CONTAINER_NAME
-	self._slotContainer.Archivable = false
-	self._slotContainer.Parent = self._obj
-
-	self._slotMap = {}
-	self._ephemeralStores = {}
-	self._transferableEphemeralKeys = {}
-	self._codeGenerator = SaveSlotCodeUtils.generateDefaultCode
-
 	self._summaryProviders = self._maid:Add(ObservableMap.new())
 
-	self._loadPromise = self._maid:GivePromise(self:_promiseLoadSlots())
+	-- UserId is read defensively because it is not indexable on every stand-in a binder can be bound to.
+	local okUserId, userId = pcall(function()
+		return self._obj.UserId
+	end)
+
+	-- Annotated rather than inlined into the call: inferring this table against the class's
+	-- intersection type tips luau-lsp into "Code is too complex to typecheck".
+	local options: HasSaveSlotsDataStore.HasSaveSlotsDataStoreOptions = {
+		ActiveSlotId = self.ActiveSlotId,
+		LastActiveSlotId = self.LastActiveSlotId,
+		ActiveTransferableEphemeralKey = self.ActiveTransferableEphemeralKey,
+		MaxSlotCount = self.MaxSlotCount,
+		SharedDataStoreService = self._sharedSaveSlotDataStoreService,
+		PromisePreSelect = function(slotId: SaveSlotData.SlotId)
+			return self:_promisePreSelectFromSaveSlotService(slotId)
+		end,
+		UserId = if okUserId and type(userId) == "number" then userId else nil,
+		UserName = self._obj.Name,
+		TrackPlaytime = true,
+	}
+
+	self._slotsDataStore =
+		self._maid:Add(HasSaveSlotsDataStore.new(self._playerDataStoreService:PromiseDataStore(self._obj), options))
+
+	-- Replicate the slot roster from the player, which is where the client reads it (see
+	-- SaveSlotDataService). The slot system itself has no idea a player is involved.
+	self._maid:GiveTask(self._slotsDataStore.SlotContainerParent:Mount(self._obj))
 
 	self._remoting = self._maid:Add(Remoting.Server.new(self._obj, "HasSaveSlots"))
 
 	self:_setupSummary()
-	self:_setupPlaytimeTracking()
 	self:_setupRemotes()
 
 	self._maid:GiveTask(HasSaveSlotsInterface.Server:Implement(self._obj, self))
@@ -121,52 +144,49 @@ function HasSaveSlots.new(player: Player, serviceBag: ServiceBag.ServiceBag): Ha
 end
 
 --[=[
+	Returns the underlying player-independent slot system, a [HasSaveSlotsDataStore]. Prefer the
+	delegating methods on this class; this is for code that needs to hand the slot system to something
+	that does not know about players.
+
+	Typed `any` deliberately: naming `HasSaveSlotsDataStore.HasSaveSlotsDataStore` anywhere in this
+	file -- even once -- pushes the old solver over its complexity budget ("Code is too complex"). The
+	delegating methods below still carry their own precise return types, which is what callers read;
+	require the module directly for a typed handle.
+
+	@return HasSaveSlotsDataStore
+]=]
+function HasSaveSlots.GetSlotsDataStore(self: HasSaveSlots): any
+	return self._slotsDataStore
+end
+
+--[=[
 	Observes the [DataStoreStage] for the active slot as a [Brio]
 ]=]
 function HasSaveSlots.ObserveActiveSlotStoreBrio(
 	self: HasSaveSlots
 ): Observable.Observable<Brio.Brio<DataStoreStage.DataStoreStage>>
-	return Rx.fromPromise(self._loadPromise):Pipe({
-		Rx.switchMap(function()
-			return self.ActiveSlotId
-				:ObserveBrio(function(slotId: SaveSlotData.SlotId?)
-					return (slotId ~= nil)
-				end)
-				:Pipe({
-					RxBrioUtils.map(function(slotId: SaveSlotData.SlotId)
-						return self:_getSlotStore(slotId)
-					end) :: any,
-				}) :: any
-		end) :: any,
-	}) :: any
+	return self._slotsDataStore:ObserveActiveSlotStoreBrio()
 end
 
 --[=[
 	Returns the [DataStoreStage] for the active slot
 ]=]
 function HasSaveSlots.PromiseActiveSlotStore(self: HasSaveSlots): Promise.Promise<DataStoreStage.DataStoreStage?>
-	return (self._loadPromise :: any):Then(function()
-		if not self.ActiveSlotId.Value then
-			return (Promise :: any).resolved(nil)
-		end
-		return self:_getSlotStore(self.ActiveSlotId.Value)
-	end)
+	return self._slotsDataStore:PromiseActiveSlotStore()
 end
 
 --[=[
 	Promises that all slots have loaded
 ]=]
 function HasSaveSlots.PromiseSlotsLoaded(self: HasSaveSlots): Promise.Promise<any>
-	return self._loadPromise
+	return self._slotsDataStore:PromiseSlotsLoaded()
 end
 
 --[=[
 	Returns whether the slot with the given ID exists
 ]=]
 function HasSaveSlots.PromiseHasSlot(self: HasSaveSlots, slotId: SaveSlotData.SlotId?): Promise.Promise<boolean>
-	return (self._loadPromise :: any):Then(function()
-		return slotId and ((self._slotMap[slotId] :: Folder?) ~= nil)
-	end)
+	return self._slotsDataStore:PromiseHasSlot(slotId)
 end
 
 -- SaveSlotService owns the callbacks and runs them. Reached through the module instance rather than by
@@ -189,72 +209,16 @@ end
 	Selects the slot with the given ID
 ]=]
 function HasSaveSlots.PromiseSelectSlot(self: HasSaveSlots, slotId: SaveSlotData.SlotId): Promise.Promise<any>
-	return (self._loadPromise :: any):Then(function()
-		if slotId == self.ActiveSlotId.Value then
-			return -- Already set
-		end
-
-		local slot = self._slotMap[slotId]
-		if not slot then
-			return (Promise :: any).rejected(`Slot \{{slotId}\} not found`)
-		end
-
-		local function setSlot()
-			self.ActiveSlotId.Value = slotId
-			SaveSlotData.LastPlayedTime:Set(slot, os.time())
-		end
-
-		local function promiseSetSlot()
-			return self._maid
-				:GivePromise(self:_promisePreSelectFromSaveSlotService(slotId))
-				:Then(function(allowed: boolean)
-					if not allowed then
-						return (Promise :: any).rejected(`Slot \{{slotId}\} refused by a pre-select callback`)
-					end
-
-					return setSlot()
-				end)
-		end
-
-		-- Initialize or save and switch
-		if self.ActiveSlotId.Value == nil then
-			return promiseSetSlot()
-		end
-
-		-- Leaving an ephemeral slot has nothing to persist, so switch without a datastore flush (the flush
-		-- exists to save the outgoing slot's progress, and an ephemeral slot has none).
-		if self:_isEphemeral(self.ActiveSlotId.Value) then
-			return promiseSetSlot()
-		end
-
-		return self._maid:GivePromise(self._dataStore:Save()):Then(promiseSetSlot)
-	end)
+	return self._slotsDataStore:PromiseSelectSlot(slotId)
 end
 
 --[=[
 	Clears the active slot selection, returning the player to a no-slot state --
 	the counterpart to [HasSaveSlots.PromiseSelectSlot], backing a "back to menu"
-	affordance. The active slot's progress is flushed first (mirroring the save
-	PromiseSelectSlot runs when switching away), and the last-active slot is
-	remembered, so [HasSaveSlots.PromiseSelectLastSaveSlot] can resume it later.
-	A no-op when no slot is active.
+	affordance. See [HasSaveSlotsDataStore.PromiseDeselectSlot].
 ]=]
 function HasSaveSlots.PromiseDeselectSlot(self: HasSaveSlots): Promise.Promise<()>
-	return (self._loadPromise :: any):Then(function()
-		if self.ActiveSlotId.Value == nil then
-			return -- Already deselected
-		end
-
-		-- An ephemeral slot has nothing to flush, so clear it without a datastore save.
-		if self:_isEphemeral(self.ActiveSlotId.Value) then
-			self.ActiveSlotId.Value = nil
-			return
-		end
-
-		return self._dataStore:Save():Then(function()
-			self.ActiveSlotId.Value = nil
-		end)
-	end)
+	return self._slotsDataStore:PromiseDeselectSlot()
 end
 
 --[=[
@@ -265,46 +229,12 @@ function HasSaveSlots.PromiseCreateSlot(
 	slotIndex: number,
 	metadata: SaveSlotCreateMetadata?
 ): Promise.Promise<SaveSlotData.SlotId>
-	return (self._loadPromise :: any):Then(function()
-		if (slotIndex < 1) or (slotIndex > self.MaxSlotCount.Value) then
-			return (Promise :: any).rejected(`Index must be in range [1, {self.MaxSlotCount.Value}]`)
-		end
-
-		for existingSlotId, slot in self._slotMap do
-			if self:_isEphemeral(existingSlotId) then
-				continue -- ephemeral slots carry no meaningful index; never let one block a real index
-			end
-			if slotIndex == SaveSlotData.SlotIndex:Get(slot) then
-				return (Promise :: any).rejected(`Slot {slotIndex} already exists`)
-			end
-		end
-
-		local slotId = HttpService:GenerateGUID(false)
-		local data = {
-			SlotId = slotId,
-			SlotIndex = slotIndex,
-			SlotName = (metadata and metadata.SlotName) or `Slot {slotIndex}`,
-			CreatedTime = os.time(),
-			Summary = metadata and metadata.Summary,
-			TimePlayed = metadata and metadata.TimePlayed,
-		}
-
-		self:_buildSlot(slotId, data, true)
-		return slotId
-	end)
+	return self._slotsDataStore:PromiseCreateSlot(slotIndex, metadata)
 end
 
 --[=[
-	Exports a slot's saved data into a plain, serializable [SaveSlotExportUtils.SaveSlotExport].
-	Rejects the main/default slot by default: its store is the player's shared root datastore, so
-	exporting it would leak the SaveSlots system data and universe-scoped global data living
-	alongside it. Only isolated non-main slot substores are exportable.
-
-	`allowMainSlot` opts out of that refusal for trusted admin tooling (see [SaveSlotCmdrService]).
-	The SaveSlots system data is always stripped from the result, so the export never carries the
-	slot roster or the other slots' saved data. Universe-scoped global data sharing the root store
-	is indistinguishable from the main slot's own data and is still carried, which is why this is
-	not exposed through [SaveSlotService].
+	Exports a slot's saved data into a plain, serializable [SaveSlotExportUtils.SaveSlotExport]. Refuses
+	the main slot unless `allowMainSlot` is set -- see [HasSaveSlotsDataStore.PromiseExportSlot].
 
 	@param slotId SlotId
 	@param allowMainSlot boolean? -- defaults to false
@@ -315,43 +245,12 @@ function HasSaveSlots.PromiseExportSlot(
 	slotId: SaveSlotData.SlotId,
 	allowMainSlot: boolean?
 ): Promise.Promise<SaveSlotExportUtils.SaveSlotExport>
-	return (self._loadPromise :: any):Then(function()
-		local slot = self._slotMap[slotId]
-		if not slot then
-			return (Promise :: any).rejected(`Slot \{{slotId}\} not found`)
-		end
-
-		local isMainSlot = SaveSlotExportUtils.isMainSlotIndex(SaveSlotData.SlotIndex:Get(slot))
-		if isMainSlot and not allowMainSlot then
-			return (Promise :: any).rejected("Cannot export the main slot")
-		end
-
-		-- As in PromiseDuplicateSlot: bring the live session's accrued time into TimePlayed before it is
-		-- read, so an export of the active slot carries its playtime up to this moment.
-		self:_flushPlaytime()
-
-		local metadata = SaveSlotData:Get(slot)
-		return self:_getSlotStore(slotId):LoadAll({}):Then(function(sourceData)
-			local data = if type(sourceData) == "table" then table.clone(sourceData) else {}
-
-			if isMainSlot then
-				-- The main slot's store is the player's shared root, so LoadAll pulls the SaveSlots system
-				-- data with it -- the slot roster and every non-main slot's saved data. Drop it so the
-				-- export is only this slot, and so importing it can't nest a whole roster inside a slot
-				-- (mirrors PromiseDuplicateSlot, which strips the same key for the same reason).
-				data[SaveSlotConstants.SYSTEM_STORE_KEY] = nil
-			end
-
-			return SaveSlotExportUtils.create(data, metadata.SlotName, metadata.Summary, metadata.TimePlayed)
-		end)
-	end)
+	return self._slotsDataStore:PromiseExportSlot(slotId, allowMainSlot)
 end
 
 --[=[
-	Imports an exported slot into a fresh slot at the lowest free non-main index, seeding the new
-	slot's store with the exported data. Never uses the main/default index -- importing onto the
-	shared root store would wipe the player's global data. Resolves to the new slot's id. Rejects a
-	malformed export, or when no non-main index is free.
+	Imports an exported slot into a fresh slot at the lowest free non-main index. See
+	[HasSaveSlotsDataStore.PromiseImportSlot].
 
 	@param export SaveSlotExportUtils.SaveSlotExport
 	@return Promise<SlotId>
@@ -360,52 +259,16 @@ function HasSaveSlots.PromiseImportSlot(
 	self: HasSaveSlots,
 	export: SaveSlotExportUtils.SaveSlotExport
 ): Promise.Promise<SaveSlotData.SlotId>
-	return (self._loadPromise :: any):Then(function()
-		if not SaveSlotExportUtils.isSaveSlotExport(export) then
-			return (Promise :: any).rejected("Bad save slot export")
-		end
-
-		-- Lowest free index strictly above the main slot: an imported slot must never occupy the
-		-- default index, whose store is the shared root datastore.
-		local usedIndices = {}
-		for existingSlotId, slot in self._slotMap do
-			if not self:_isEphemeral(existingSlotId) then
-				usedIndices[SaveSlotData.SlotIndex:Get(slot)] = true
-			end
-		end
-		local freeIndex = SaveSlotConstants.DEFAULT_SLOT_INDEX + 1
-		while usedIndices[freeIndex] do
-			freeIndex += 1
-		end
-		if freeIndex > self.MaxSlotCount.Value then
-			return (Promise :: any).rejected("No free non-main slot index available")
-		end
-
-		return self:PromiseCreateSlot(freeIndex, {
-			SlotName = export.slotName,
-			Summary = export.summary,
-			-- Absent on exports taken before playtime was carried; those import as before, with no playtime.
-			TimePlayed = export.timePlayed,
-		}):Then(function(newSlotId: SaveSlotData.SlotId)
-			-- freeIndex is always > DEFAULT_SLOT_INDEX, so this store is an isolated substore
-			-- (never the shared root); a plain Overwrite cannot touch system or global data.
-			self:_getSlotStore(newSlotId):Overwrite(export.data)
-
-			-- Flush so the imported slot survives a crash before the next autosave.
-			return self._dataStore:Save():Then(function()
-				return newSlotId
-			end)
-		end)
-	end)
+	return self._slotsDataStore:PromiseImportSlot(export)
 end
 
 --[=[
-	Exports a non-main slot (see [HasSaveSlots.PromiseExportSlot]) and writes it to the shared save
-	slot store under the given key.
+	Exports a non-main slot and writes it to the shared save slot store under the given key. See
+	[HasSaveSlotsDataStore.PromiseSaveSlotToSharedDataStore].
 
 	@param slotId SlotId
 	@param key string
-	@param allowMainSlot boolean? -- defaults to false, see [HasSaveSlots.PromiseExportSlot]
+	@param allowMainSlot boolean? -- defaults to false
 	@return Promise<boolean>
 ]=]
 function HasSaveSlots.PromiseSaveSlotToSharedDataStore(
@@ -414,14 +277,12 @@ function HasSaveSlots.PromiseSaveSlotToSharedDataStore(
 	key: string,
 	allowMainSlot: boolean?
 ): Promise.Promise<boolean>
-	return self:PromiseExportSlot(slotId, allowMainSlot):Then(function(export)
-		return self._sharedSaveSlotDataStoreService:PromiseWrite(key, export)
-	end)
+	return self._slotsDataStore:PromiseSaveSlotToSharedDataStore(slotId, key, allowMainSlot)
 end
 
 --[=[
-	Reads an export from the shared save slot store and imports it into a fresh non-main slot (see
-	[HasSaveSlots.PromiseImportSlot]). Rejects when no export is stored under the key.
+	Reads an export from the shared save slot store and imports it into a fresh non-main slot. See
+	[HasSaveSlotsDataStore.PromiseImportSlotFromSharedDataStore].
 
 	@param key string
 	@return Promise<SlotId>
@@ -430,25 +291,12 @@ function HasSaveSlots.PromiseImportSlotFromSharedDataStore(
 	self: HasSaveSlots,
 	key: string
 ): Promise.Promise<SaveSlotData.SlotId>
-	-- Maid-owned like the other shared-store reads: this one is not even gated behind _loadPromise, so
-	-- nothing upstream rejects it when the player leaves mid-read.
-	return self._maid:GivePromise(self._sharedSaveSlotDataStoreService:PromiseRead(key)):Then(function(export)
-		if not self.Destroy then
-			return (Promise :: any).rejected() -- Destroyed. Empty, so it stays out of the logs
-		end
-
-		if export == nil then
-			return (Promise :: any).rejected(`No save slot stored under \{{key}\}`)
-		end
-		return self:PromiseImportSlot(export)
-	end)
+	return self._slotsDataStore:PromiseImportSlotFromSharedDataStore(key)
 end
 
 --[=[
-	Loads the export stored under the given shared-store key into a fresh ephemeral slot and selects it,
-	remembering the key so a teleport can carry the slot forward (see the transferable-ephemeral teleport
-	provider in [SaveSlotService]). Like every ephemeral slot it is never persisted, stays out of the slot
-	list, and is torn down on deselect. Rejects when no valid export is stored under the key.
+	Loads the export stored under the given shared-store key into a fresh ephemeral slot and selects it.
+	See [HasSaveSlotsDataStore.PromiseSelectTransferableEphemeralSlot].
 
 	@param key string
 	@return Promise<SlotId>
@@ -457,75 +305,17 @@ function HasSaveSlots.PromiseSelectTransferableEphemeralSlot(
 	self: HasSaveSlots,
 	key: string
 ): Promise.Promise<SaveSlotData.SlotId>
-	return (self._loadPromise :: any):Then(function()
-		-- Maid-owned for the same reason as _promiseLoadSlots: this read runs on the join path (a teleport
-		-- arriving with an ephemeral key) and outlives a leave, and its continuation builds a slot.
-		return self._maid:GivePromise(self._sharedSaveSlotDataStoreService:PromiseRead(key)):Then(function(export)
-			if not self.Destroy then
-				-- Destroyed. Rejecting keeps the Promise<SlotId> contract honest -- a nil fulfilment reads as
-				-- "imported, no slot" to a caller like SaveSlotCmdrService. Empty, so it stays out of the logs.
-				return (Promise :: any).rejected()
-			end
-
-			if export == nil then
-				return (Promise :: any).rejected(`No save slot stored under \{{key}\}`)
-			end
-			if not SaveSlotExportUtils.isSaveSlotExport(export) then
-				return (Promise :: any).rejected("Bad save slot export in shared store")
-			end
-
-			local slotId = HttpService:GenerateGUID(false)
-			self:_buildSlot(slotId, {
-				SlotId = slotId,
-				SlotIndex = SaveSlotConstants.EPHEMERAL_SLOT_INDEX,
-				SlotName = export.slotName or "Transfer",
-				CreatedTime = os.time(),
-				Summary = export.summary,
-			}, true, true)
-
-			-- Seed the in-memory store before selecting so a consumer of ObserveActiveSlotStoreBrio never
-			-- observes an empty active slot; then remember the key this slot transfers under.
-			self:_getSlotStore(slotId):Overwrite(export.data)
-			self._transferableEphemeralKeys[slotId] = key
-
-			return self:PromiseSelectSlot(slotId):Then(function()
-				return slotId
-			end)
-		end)
-	end)
+	return self._slotsDataStore:PromiseSelectTransferableEphemeralSlot(key)
 end
 
 --[=[
-	Builds this player's teleport slice for a transferable ephemeral slot: re-saves the active slot's
-	*current live* data to the shared store under its key and returns a slice carrying that key. Resolves
-	nil when the active slot is not a transferable ephemeral slot. A failed re-save degrades to nil so a
-	teleport is never blocked (the destination then re-loads the last saved state). Asynchronous -- it is
-	consumed through [TeleportDataService.PromiseBuildTeleportData].
+	Builds this player's teleport slice for a transferable ephemeral slot. See
+	[HasSaveSlotsDataStore.PromiseBuildEphemeralTransferSlice].
 
 	@return Promise<{ [string]: any }?>
 ]=]
 function HasSaveSlots.PromiseBuildEphemeralTransferSlice(self: HasSaveSlots): Promise.Promise<{ [string]: any }?>
-	return (self._loadPromise :: any):Then(function(): any
-		local slotId = self.ActiveSlotId.Value
-		if not slotId then
-			return nil
-		end
-		local storeKey = self._transferableEphemeralKeys[slotId]
-		if not storeKey then
-			return nil
-		end
-
-		return self:PromiseExportSlot(slotId)
-			:Then(function(export)
-				return self._sharedSaveSlotDataStoreService:PromiseWrite(storeKey, export)
-			end)
-			:Then(function()
-				return { [SaveSlotConstants.TELEPORT_DATA_EPHEMERAL_KEY] = storeKey }
-			end)
-			:Catch(function()
-				return nil
-			end)
-	end)
+	return self._slotsDataStore:PromiseBuildEphemeralTransferSlice()
 end
 
 --[=[
@@ -562,36 +352,15 @@ end
 	@param generator CodeGenerator
 ]=]
 function HasSaveSlots.SetCodeGenerator(self: HasSaveSlots, generator: SaveSlotCodeUtils.CodeGenerator): ()
-	assert(type(generator) == "function", "Bad generator")
-	self._codeGenerator = generator
-end
-
--- Builds a code for a slot from the configured generator, resolving the owner's identity defensively
--- (player.UserId is not indexable on every stand-in) and the slot's name/index from its metadata.
-function HasSaveSlots._generateCode(self: HasSaveSlots, slotId: SaveSlotData.SlotId): string
-	local slot = self._slotMap[slotId]
-	local metadata = if slot then SaveSlotData:Get(slot) else nil
-	local okUserId, userId = pcall(function()
-		return self._obj.UserId
-	end)
-
-	return self._codeGenerator({
-		userId = if okUserId and type(userId) == "number" then userId else nil,
-		userName = self._obj.Name,
-		slotName = metadata and metadata.SlotName,
-		slotIndex = metadata and metadata.SlotIndex,
-	})
+	self._slotsDataStore:SetCodeGenerator(generator)
 end
 
 --[=[
-	Exports a slot to the shared store under a fresh generated code and resolves to that code. The code
-	is a shareable handle other sessions load with [HasSaveSlots.PromiseImportEphemeralSaveSlotFromCode].
-	Defaults to the active slot. Refuses the main slot unless `allowMainSlot` is set (see
-	[HasSaveSlots.PromiseExportSlot]). The code format comes from the configured generator (see
-	[HasSaveSlots.SetCodeGenerator]).
+	Exports a slot to the shared store under a fresh generated code and resolves to that code. See
+	[HasSaveSlotsDataStore.PromiseExportSaveSlotToCode].
 
 	@param slotId SlotId? -- defaults to the active slot
-	@param allowMainSlot boolean? -- defaults to false, see [HasSaveSlots.PromiseExportSlot]
+	@param allowMainSlot boolean? -- defaults to false
 	@return Promise<string>
 ]=]
 function HasSaveSlots.PromiseExportSaveSlotToCode(
@@ -599,22 +368,12 @@ function HasSaveSlots.PromiseExportSaveSlotToCode(
 	slotId: SaveSlotData.SlotId?,
 	allowMainSlot: boolean?
 ): Promise.Promise<string>
-	return (self._loadPromise :: any):Then(function()
-		local targetSlotId = slotId or self.ActiveSlotId.Value
-		if not targetSlotId then
-			return (Promise :: any).rejected("No slot to export")
-		end
-
-		local code = self:_generateCode(targetSlotId)
-		return self:PromiseSaveSlotToSharedDataStore(targetSlotId, code, allowMainSlot):Then(function()
-			return code
-		end)
-	end)
+	return self._slotsDataStore:PromiseExportSaveSlotToCode(slotId, allowMainSlot)
 end
 
 --[=[
-	Loads the slot stored under the given code into a fresh transferable ephemeral slot and selects it
-	(see [HasSaveSlots.PromiseSelectTransferableEphemeralSlot]). Resolves to the new slot id.
+	Loads the slot stored under the given code into a fresh transferable ephemeral slot and selects it.
+	See [HasSaveSlotsDataStore.PromiseImportEphemeralSaveSlotFromCode].
 
 	@param code string
 	@return Promise<SlotId>
@@ -623,16 +382,15 @@ function HasSaveSlots.PromiseImportEphemeralSaveSlotFromCode(
 	self: HasSaveSlots,
 	code: string
 ): Promise.Promise<SaveSlotData.SlotId>
-	return self:PromiseSelectTransferableEphemeralSlot(code)
+	return self._slotsDataStore:PromiseImportEphemeralSaveSlotFromCode(code)
 end
 
 --[=[
-	Exports a slot as a raw JSON string (no shared store), for direct inspection or attaching to a bug
-	report. Defaults to the active slot. Refuses the main slot unless `allowMainSlot` is set (see
-	[HasSaveSlots.PromiseExportSlot]).
+	Exports a slot as a raw JSON string (no shared store). See
+	[HasSaveSlotsDataStore.PromiseExportSaveSlotToJson].
 
 	@param slotId SlotId? -- defaults to the active slot
-	@param allowMainSlot boolean? -- defaults to false, see [HasSaveSlots.PromiseExportSlot]
+	@param allowMainSlot boolean? -- defaults to false
 	@return Promise<string>
 ]=]
 function HasSaveSlots.PromiseExportSaveSlotToJson(
@@ -640,110 +398,23 @@ function HasSaveSlots.PromiseExportSaveSlotToJson(
 	slotId: SaveSlotData.SlotId?,
 	allowMainSlot: boolean?
 ): Promise.Promise<string>
-	return (self._loadPromise :: any):Then(function()
-		local targetSlotId = slotId or self.ActiveSlotId.Value
-		if not targetSlotId then
-			return (Promise :: any).rejected("No slot to export")
-		end
-
-		return self:PromiseExportSlot(targetSlotId, allowMainSlot):Then(function(export)
-			return HttpService:JSONEncode(export)
-		end)
-	end)
+	return self._slotsDataStore:PromiseExportSaveSlotToJson(slotId, allowMainSlot)
 end
 
 --[=[
-	Duplicates the slot with the given ID into a new slot at the lowest free index,
-	copying its saved data and accrued playtime. Resolves to the new slot's id. The copy is not
-	selected, its timestamps and session counters start fresh, and its name is suffixed with
-	" (Copy)". Rejects when the source slot is missing or every index is in use.
-
-	An ephemeral slot may be duplicated: the copy is a real, persisted slot seeded with the ephemeral
-	slot's live in-memory data, which is how a throwaway session is turned into a save. It keeps the
-	source's name unsuffixed (the copy is the first real slot for that session, not a second copy of an
-	existing save) and, like any duplicate, is not selected -- see
-	[HasSaveSlots.PromisePersistEphemeralSlot] for the version that continues play on the new slot.
+	Duplicates the slot with the given ID into a new slot at the lowest free index, copying its saved
+	data and accrued playtime. See [HasSaveSlotsDataStore.PromiseDuplicateSlot].
 ]=]
 function HasSaveSlots.PromiseDuplicateSlot(
 	self: HasSaveSlots,
 	slotId: SaveSlotData.SlotId
 ): Promise.Promise<SaveSlotData.SlotId>
-	return (self._loadPromise :: any):Then(function()
-		local sourceSlot = self._slotMap[slotId]
-		if not sourceSlot then
-			return (Promise :: any).rejected(`Slot \{{slotId}\} not found`)
-		end
-
-		local sourceIsEphemeral = self:_isEphemeral(slotId)
-
-		-- The source is usually the live active slot, whose current session has not landed in TimePlayed
-		-- yet; fold it in first so the copy carries the playtime the player is actually looking at.
-		self:_flushPlaytime()
-
-		-- Lowest free positive index, filling gaps left by deletions (mirrors PromiseSelectNewSaveSlot).
-		local usedIndices = {}
-		for existingSlotId, slot in self._slotMap do
-			if not self:_isEphemeral(existingSlotId) then
-				usedIndices[SaveSlotData.SlotIndex:Get(slot)] = true
-			end
-		end
-		local freeIndex = 1
-		while usedIndices[freeIndex] do
-			freeIndex += 1
-		end
-		if freeIndex > self.MaxSlotCount.Value then
-			return (Promise :: any).rejected("All slots are already in use")
-		end
-
-		local sourceMetadata = SaveSlotData:Get(sourceSlot)
-
-		-- Read the source's saved data before creating the copy so a read failure leaves no orphan slot.
-		return self:_getSlotStore(slotId):LoadAll({}):Then(function(sourceData)
-			-- The default slot shares the player's root store with the SaveSlots system data. Never carry
-			-- that system key across into the copy's saved data.
-			local slotData = if type(sourceData) == "table" then table.clone(sourceData) else {}
-			slotData[SaveSlotConstants.SYSTEM_STORE_KEY] = nil
-
-			return self
-				:PromiseCreateSlot(freeIndex, {
-					SlotName = if sourceIsEphemeral
-						then sourceMetadata.SlotName
-						else `{sourceMetadata.SlotName} (Copy)`,
-					Summary = sourceMetadata.Summary,
-					-- Playtime describes the progress being copied, not the slot it happens to live in, so
-					-- it travels with that progress rather than resetting the copy to a never-played slot.
-					TimePlayed = sourceMetadata.TimePlayed,
-				})
-				:Then(function(newSlotId: SaveSlotData.SlotId)
-					local destStore = self:_getSlotStore(newSlotId)
-
-					if destStore == self._dataStore then
-						-- The copy is the default slot, whose store is the shared root. Merge so the SaveSlots
-						-- system store living alongside it survives (a plain Overwrite would wipe it).
-						destStore:OverwriteMerge(slotData)
-					else
-						destStore:Overwrite(slotData)
-					end
-
-					-- Flush so the duplicated data survives a crash before the next autosave.
-					return self._dataStore:Save():Then(function()
-						return newSlotId
-					end)
-				end)
-		end)
-	end)
+	return self._slotsDataStore:PromiseDuplicateSlot(slotId)
 end
 
 --[=[
-	Turns an ephemeral slot into a real save: duplicates it into a fresh persisted slot at the lowest free
-	index (see [HasSaveSlots.PromiseDuplicateSlot]) and selects that slot, so play continues on data that
-	is now being written. Defaults to the active slot, which in practice is the only ephemeral slot there
-	is -- one stops existing the moment it stops being active. Resolves to the new slot's id.
-
-	The copy is taken from the ephemeral slot's live in-memory data at the moment this runs, and selecting
-	the new slot retires the ephemeral one, so anything written to the old store between the copy and the
-	selection is dropped. Rejects when no slot is active, the slot is missing, the slot is not ephemeral,
-	or every index is in use.
+	Turns an ephemeral slot into a real save and selects it. See
+	[HasSaveSlotsDataStore.PromisePersistEphemeralSlot].
 
 	@param slotId SlotId? -- defaults to the active slot
 	@return Promise<SlotId>
@@ -752,88 +423,14 @@ function HasSaveSlots.PromisePersistEphemeralSlot(
 	self: HasSaveSlots,
 	slotId: SaveSlotData.SlotId?
 ): Promise.Promise<SaveSlotData.SlotId>
-	return (self._loadPromise :: any):Then(function()
-		local targetSlotId = slotId or self.ActiveSlotId.Value
-		if not targetSlotId then
-			return (Promise :: any).rejected("No slot to persist")
-		end
-
-		if not self._slotMap[targetSlotId] then
-			return (Promise :: any).rejected(`Slot \{{targetSlotId}\} not found`)
-		end
-
-		if not self:_isEphemeral(targetSlotId) then
-			return (Promise :: any).rejected(`Slot \{{targetSlotId}\} is already persisted`)
-		end
-
-		return self:PromiseDuplicateSlot(targetSlotId):Then(function(newSlotId: SaveSlotData.SlotId)
-			return self:PromiseSelectSlot(newSlotId):Then(function()
-				return newSlotId
-			end)
-		end)
-	end)
+	return self._slotsDataStore:PromisePersistEphemeralSlot(slotId)
 end
 
 --[=[
-	Deletes the slot with the given ID. Deleting an ephemeral slot ends that session -- it is deselected
-	and retired along with its in-memory store -- rather than being refused for being the active slot.
+	Deletes the slot with the given ID. See [HasSaveSlotsDataStore.PromiseDeleteSlot].
 ]=]
 function HasSaveSlots.PromiseDeleteSlot(self: HasSaveSlots, slotId: SaveSlotData.SlotId): Promise.Promise<any>
-	return (self._loadPromise :: any):Then(function()
-		-- An ephemeral slot only exists while it is active, so deleting one is retiring it. Nothing was
-		-- ever persisted, so there is no stored data to wipe afterwards -- hence the early return.
-		if self:_isEphemeral(slotId) then
-			if slotId == self.ActiveSlotId.Value then
-				return self:PromiseDeselectSlot()
-			end
-			self:_destroyEphemeralSlot(slotId)
-			return (Promise :: any).resolved()
-		end
-
-		if slotId == self.ActiveSlotId.Value then
-			return (Promise :: any).rejected("Cannot delete active slot")
-		end
-
-		local slot = self._slotMap[slotId]
-		if not slot then
-			return (Promise :: any).rejected(`Slot \{{slotId}\} not found`)
-		end
-
-		self._maid[slotId] = nil
-
-		-- The continue pointer must not outlive its slot. When the deleted slot is the one the player
-		-- would resume on (e.g. the just-deselected active slot, or a never-reselected last-active),
-		-- clear the last-active memory so "Continue" stops offering a slot that no longer exists.
-		if slotId == self._lastActiveSlotId then
-			self._lastActiveSlotId = nil
-			self.LastActiveSlotId.Value = nil
-		end
-
-		-- Wipe default slot
-		local slotIndex = SaveSlotData.SlotIndex:Get(slot)
-		local deletePromise = nil
-
-		if slotIndex == SaveSlotConstants.DEFAULT_SLOT_INDEX then
-			deletePromise = self._dataStore:PromiseKeyList():Then(function(keys)
-				for _, key in keys do
-					if key ~= SaveSlotConstants.SYSTEM_STORE_KEY then
-						self._dataStore:Delete(key)
-					end
-				end
-				self._metadataStore:Delete(slotId)
-			end)
-		else
-			-- Or delete slot from substore
-			self._systemStore:GetSubStore(SaveSlotConstants.SLOT_STORE_KEY):Delete(slotId)
-			self._metadataStore:Delete(slotId)
-			deletePromise = (Promise :: any).resolved()
-		end
-
-		-- Flush the deletion to prevent stale reads
-		return deletePromise:Then(function()
-			return self._dataStore:Save()
-		end)
-	end)
+	return self._slotsDataStore:PromiseDeleteSlot(slotId)
 end
 
 --[=[
@@ -842,104 +439,25 @@ end
 	are gone.
 ]=]
 function HasSaveSlots.PromiseDeleteAllSlots(self: HasSaveSlots): Promise.Promise<any>
-	return (self._loadPromise :: any):Then(function()
-		-- Clear the selection first so the previously active slot is deletable
-		self.ActiveSlotId.Value = nil
-		self._lastActiveSlotId = nil
-		self.LastActiveSlotId.Value = nil
-
-		local slotIds = {}
-		for slotId in self._slotMap do
-			table.insert(slotIds, slotId)
-		end
-
-		-- Delete sequentially to avoid concurrent datastore saves
-		local promise = (Promise :: any).resolved()
-		for _, slotId in slotIds do
-			promise = promise:Then(function()
-				return self:PromiseDeleteSlot(slotId)
-			end)
-		end
-		return promise
-	end)
+	return self._slotsDataStore:PromiseDeleteAllSlots()
 end
 
 --[=[
-	Resets the slot with the given id to a fresh empty one -- equivalent to deleting the slot
-	and creating a new one at the same index. The slot keeps its index and name; its saved data
-	and metadata (timestamps) start fresh. Resolves to the new slot id.
-
-	When the reset slot is the active slot, the selection clears and then reselects the fresh
-	slot: everything bound to [HasSaveSlots.ObserveActiveSlotStoreBrio] tears down as the
-	selection clears and rebuilds against the empty store on reselect, so consumers reset
-	reactively without wiping their own state. A non-active slot is left unselected, and its
-	"Continue" pointer (when it was the last-active slot) is carried across to the fresh id so
-	the reset slot stays resumable. Rejects when the slot is missing.
+	Resets the slot with the given id to a fresh empty one. See [HasSaveSlotsDataStore.PromiseResetSlot].
 ]=]
 function HasSaveSlots.PromiseResetSlot(
 	self: HasSaveSlots,
 	slotId: SaveSlotData.SlotId
 ): Promise.Promise<SaveSlotData.SlotId>
-	return (self._loadPromise :: any):Then(function(): any
-		local slot = self._slotMap[slotId]
-		if not slot then
-			return (Promise :: any).rejected(`Slot \{{slotId}\} not found`)
-		end
-
-		if self:_isEphemeral(slotId) then
-			return (Promise :: any).rejected("Cannot reset an ephemeral slot")
-		end
-
-		local metadata = SaveSlotData:Get(slot)
-		local wasActive = (slotId == self.ActiveSlotId.Value)
-		-- PromiseDeleteSlot clears the continue pointer when it targets the last-active slot; capture it
-		-- now so the non-active path can move it onto the fresh slot id below.
-		local wasLastActive = (slotId == self._lastActiveSlotId)
-
-		if wasActive then
-			-- Clear the selection so the slot is deletable, skipping the deselect flush that would
-			-- persist progress we are about to delete.
-			self.ActiveSlotId.Value = nil
-		end
-
-		return self:PromiseDeleteSlot(slotId)
-			:Then(function()
-				return self:PromiseCreateSlot(metadata.SlotIndex, { SlotName = metadata.SlotName })
-			end)
-			:Then(function(newSlotId: SaveSlotData.SlotId)
-				if wasActive then
-					-- Reselecting restores the continue pointer via the ActiveSlotId hook.
-					return self:PromiseSelectSlot(newSlotId):Then(function()
-						return newSlotId
-					end)
-				end
-
-				-- Non-active reset never reselects, so carry the resume pointer onto the fresh id
-				-- ourselves when this slot was the one "Continue" would resume.
-				if wasLastActive then
-					self._lastActiveSlotId = newSlotId
-					self.LastActiveSlotId.Value = newSlotId
-				end
-
-				return newSlotId
-			end)
-	end)
+	return self._slotsDataStore:PromiseResetSlot(slotId)
 end
 
 --[=[
-	Resets the active slot to a fresh empty one -- see [HasSaveSlots.PromiseResetSlot]. The slot
-	keeps its index and name; its saved data and metadata (timestamps) start fresh, and the fresh
-	slot stays selected. Resolves to the new slot id, or nil when no slot is active.
+	Resets the active slot to a fresh empty one -- see [HasSaveSlots.PromiseResetSlot]. Resolves to the
+	new slot id, or nil when no slot is active.
 ]=]
 function HasSaveSlots.PromiseResetActiveSlot(self: HasSaveSlots): Promise.Promise<SaveSlotData.SlotId?>
-	return (self._loadPromise :: any):Then(function(): any
-		local slotId = self.ActiveSlotId.Value
-		if not slotId then
-			return nil -- Nothing selected to reset
-		end
-
-		return self:PromiseResetSlot(slotId)
-	end)
+	return self._slotsDataStore:PromiseResetActiveSlot()
 end
 
 --[=[
@@ -950,20 +468,7 @@ function HasSaveSlots.PromiseSetSlotMetadata(
 	slotId: SaveSlotData.SlotId,
 	data: SaveSlotData.SaveSlotMetadata
 ): Promise.Promise<any>
-	if data.SlotId and (data.SlotId ~= slotId) then
-		return (Promise :: any).rejected("SlotId is locked")
-	end
-
-	return (self._loadPromise :: any):Then(function()
-		local slot = self._slotMap[slotId]
-
-		-- Routing depends on immutable indices to distinguish the default slot
-		if data.SlotIndex and (data.SlotIndex ~= SaveSlotData.SlotIndex:Get(slot)) then
-			return (Promise :: any).rejected("SlotIndex is locked")
-		end
-
-		SaveSlotData:Set(slot, data)
-	end)
+	return self._slotsDataStore:PromiseSetSlotMetadata(slotId, data)
 end
 
 --[=[
@@ -973,10 +478,7 @@ function HasSaveSlots.PromiseGetSlotMetadata(
 	self: HasSaveSlots,
 	slotId: SaveSlotData.SlotId
 ): Promise.Promise<SaveSlotData.SaveSlotMetadata?>
-	return (self._loadPromise :: any):Then(function()
-		local slot = self._slotMap[slotId]
-		return (Promise :: any).resolved(slot and SaveSlotData:Get(slot))
-	end)
+	return self._slotsDataStore:PromiseGetSlotMetadata(slotId)
 end
 
 --[=[
@@ -986,26 +488,14 @@ function HasSaveSlots.PromiseSlotIdFromIndex(
 	self: HasSaveSlots,
 	slotIndex: number
 ): Promise.Promise<SaveSlotData.SlotId?>
-	return (self._loadPromise :: any):Then(function()
-		for slotId, slot in self._slotMap do
-			if self:_isEphemeral(slotId) then
-				continue -- ephemeral slots are never addressable by index
-			end
-			if slotIndex == SaveSlotData.SlotIndex:Get(slot) then
-				return (Promise :: any).resolved(slotId)
-			end
-		end
-		return (Promise :: any).resolved(nil)
-	end)
+	return self._slotsDataStore:PromiseSlotIdFromIndex(slotIndex)
 end
 
 --[=[
 	Gets the last active slot ID
 ]=]
 function HasSaveSlots.PromiseLastActiveSlotId(self: HasSaveSlots): Promise.Promise<SaveSlotData.SlotId?>
-	return (self._loadPromise :: any):Then(function()
-		return self.ActiveSlotId.Value or self._lastActiveSlotId
-	end)
+	return self._slotsDataStore:PromiseLastActiveSlotId()
 end
 
 --[=[
@@ -1014,21 +504,7 @@ end
 	"Continue" affordance that every save-slot consumer tends to need.
 ]=]
 function HasSaveSlots.PromiseSelectLastSaveSlot(self: HasSaveSlots): Promise.Promise<SaveSlotData.SlotId?>
-	return self:PromiseLastActiveSlotId():Then(function(lastActiveSlotId: SaveSlotData.SlotId?)
-		if not lastActiveSlotId then
-			return nil
-		end
-
-		return self:PromiseHasSlot(lastActiveSlotId):Then(function(hasLastSlot: boolean)
-			if not hasLastSlot then
-				return nil
-			end
-
-			return self:PromiseSelectSlot(lastActiveSlotId):Then(function()
-				return lastActiveSlotId
-			end)
-		end)
-	end)
+	return self._slotsDataStore:PromiseSelectLastSaveSlot()
 end
 
 --[=[
@@ -1037,45 +513,12 @@ end
 	affordance.
 ]=]
 function HasSaveSlots.PromiseSelectNewSaveSlot(self: HasSaveSlots): Promise.Promise<SaveSlotData.SlotId?>
-	return (self._loadPromise :: any):Then(function(): any
-		local usedIndices = {}
-		for existingSlotId, slot in self._slotMap do
-			if not self:_isEphemeral(existingSlotId) then
-				usedIndices[SaveSlotData.SlotIndex:Get(slot)] = true
-			end
-		end
-
-		-- Lowest free positive index. Fills gaps left by deletions (delete slot 2 of
-		-- [1,2,3] and the next new slot reuses index 2). With a finite MaxSlotCount this
-		-- returns nil once [1, MaxSlotCount] is full; with an unbounded count (math.huge)
-		-- there is always a next integer, so it never returns nil.
-		local freeIndex = 1
-		while usedIndices[freeIndex] do
-			freeIndex += 1
-		end
-
-		if freeIndex > self.MaxSlotCount.Value then
-			return nil
-		end
-
-		return self:PromiseCreateSlot(freeIndex):Then(function(slotId: SaveSlotData.SlotId)
-			return self:PromiseSelectSlot(slotId):Then(function()
-				return slotId
-			end)
-		end)
-	end)
+	return self._slotsDataStore:PromiseSelectNewSaveSlot()
 end
 
 --[=[
-	Creates a fresh ephemeral slot and selects it, resolving to its id. An ephemeral slot is selectable and
-	active exactly like a real one -- it drives [HasSaveSlots.ObserveActiveSlotStoreBrio], summaries, and
-	playtime the same way, and its metadata replicates to the client like any other slot's, so UI can render
-	the active session's name and summary -- but it is never persisted: no metadata is written, its data store
-	is in-memory, and it is torn down the moment it stops being the active slot. It is also excluded from the
-	save list ([SaveSlotDataService.GetSlotList] / [SaveSlotDataService.ObserveSlotList]), so it never shows up
-	as something the player can return to. Selecting it never disturbs the persisted active-slot pointer or the
-	"Continue" target, so the real slot the player came from resumes untouched afterward. Backs a throwaway
-	session (e.g. exploring a lobby) that must leave no trace on save data.
+	Creates a fresh ephemeral slot and selects it, resolving to its id. See
+	[HasSaveSlotsDataStore.PromiseSelectEphemeralSlot].
 
 	@param metadata SaveSlotCreateMetadata? -- optional SlotName/Summary for the in-memory slot
 	@return Promise<SlotId>
@@ -1084,22 +527,7 @@ function HasSaveSlots.PromiseSelectEphemeralSlot(
 	self: HasSaveSlots,
 	metadata: SaveSlotCreateMetadata?
 ): Promise.Promise<SaveSlotData.SlotId>
-	return (self._loadPromise :: any):Then(function()
-		local slotId = HttpService:GenerateGUID(false)
-		local data = {
-			SlotId = slotId,
-			SlotIndex = SaveSlotConstants.EPHEMERAL_SLOT_INDEX,
-			SlotName = (metadata and metadata.SlotName) or "Ephemeral",
-			CreatedTime = os.time(),
-			Summary = metadata and metadata.Summary,
-		}
-
-		self:_buildSlot(slotId, data, true, true)
-
-		return self:PromiseSelectSlot(slotId):Then(function()
-			return slotId
-		end)
-	end)
+	return self._slotsDataStore:PromiseSelectEphemeralSlot(metadata)
 end
 
 --[=[
@@ -1135,198 +563,6 @@ function HasSaveSlots.PromiseIncomingSlotId(self: HasSaveSlots): Promise.Promise
 		end)
 end
 
--- Maid-owning every hop is what makes this safe, for the same reason as the selection chain in
--- SaveSlotService: a session-locked or retrying read settles long after the player left, and Promise has no
--- cancellation, so owning only the outermost promise detaches nothing upstream. Calling into this destroyed
--- (metatable-stripped) object -- or into its destroyed stores -- throws, and because a datastore load
--- resolves inside its UpdateAsync transform, that throw surfaces as a "Transform function error" and takes
--- the transform's write down with it. The liveness checks cover a settle that lands mid-teardown, before
--- this promise's own maid task was reached. See SaveSlotLateSettle.spec.
-function HasSaveSlots._promiseLoadSlots(self: HasSaveSlots): Promise.Promise<{}>
-	return self._maid:GivePromise(self._playerDataStoreService:PromiseDataStore(self._obj)):Then(function(dataStore)
-		if not self.Destroy then
-			return nil -- Destroyed
-		end
-
-		self._dataStore = dataStore
-		self._systemStore = dataStore:GetSubStore(SaveSlotConstants.SYSTEM_STORE_KEY)
-		self._metadataStore = self._systemStore:GetSubStore(SaveSlotConstants.METADATA_STORE_KEY)
-
-		return self._maid:GivePromise(self._metadataStore:LoadAll({})):Then(function(metadata)
-			if not self.Destroy then
-				return nil -- Destroyed
-			end
-
-			for slotId, data in metadata do
-				self:_buildSlot(slotId, data)
-			end
-
-			return self._maid
-				:GivePromise(self._systemStore:Load("activeSlotId"))
-				:Then(function(activeId: SaveSlotData.SlotId?)
-					if not self.Destroy then
-						return nil -- Destroyed
-					end
-
-					self._lastActiveSlotId = activeId
-					self.LastActiveSlotId.Value = activeId
-
-					-- The persisted active-slot pointer and the replicated "Continue" target only ever track real
-					-- slots. This replaces StoreOnValueChange so an ephemeral selection is invisible to both:
-					-- entering one leaves them pinned to the real slot, and the ephemeral slot is torn down the
-					-- moment it stops being active. We track the id we are leaving to know which of those to do.
-					local previousActiveSlotId: SaveSlotData.SlotId? = activeId
-
-					self._maid:GiveTask(self.ActiveSlotId.Changed:Connect(function()
-						local active = self.ActiveSlotId.Value
-
-						-- Replicate the active transferable-ephemeral slot's shared-store key (nil otherwise) so a
-						-- client-initiated teleport can carry it forward (SaveSlotServiceClient's provider reads this).
-						self.ActiveTransferableEphemeralKey.Value = if active
-							then self._transferableEphemeralKeys[active]
-							else nil
-
-						local leftSlotId = previousActiveSlotId
-						-- Read before the retire below, while the outgoing slot is still in the map.
-						local leavingEphemeral = self:_isEphemeral(leftSlotId)
-						previousActiveSlotId = active
-
-						-- Persist the pointer + remember the Continue target only for real-slot transitions
-						-- (real -> real, real -> nil deselect, nil -> real). Skip both ephemeral cases: entering an
-						-- ephemeral slot must stay invisible to persistence, and leaving one back to no slot must
-						-- leave the real pointer pinned where it was.
-						local enteringEphemeral = self:_isEphemeral(active)
-						local leavingEphemeralToMenu = leavingEphemeral and active == nil
-						if not (enteringEphemeral or leavingEphemeralToMenu) then
-							self._systemStore:Store("activeSlotId", active)
-							if active ~= nil then
-								self._lastActiveSlotId = active
-								self.LastActiveSlotId.Value = active
-							end
-						end
-
-						-- An ephemeral slot exists only while it is the active slot; retire the one we just left.
-						if leavingEphemeral and leftSlotId ~= active then
-							self:_destroyEphemeralSlot(leftSlotId :: SaveSlotData.SlotId)
-						end
-					end))
-
-					-- Matches the liveness-guard returns above, which make this callback's inferred return
-					-- type nil; falling off the end instead returns no values at all.
-					return nil
-				end)
-		end)
-	end)
-end
-
-function HasSaveSlots._getSlotStore(self: HasSaveSlots, slotId: SaveSlotData.SlotId): DataStoreStage.DataStoreStage
-	local ephemeralStore = self._ephemeralStores[slotId]
-	if ephemeralStore then
-		return ephemeralStore
-	end
-
-	local slot = self._slotMap[slotId]
-	if slot and (SaveSlotData.SlotIndex:Get(slot) == SaveSlotConstants.DEFAULT_SLOT_INDEX) then
-		return self._dataStore
-	end
-	return self._systemStore:GetSubStore(SaveSlotConstants.SLOT_STORE_KEY):GetSubStore(slotId)
-end
-
-local MUTABLE_METADATA_KEYS =
-	{ "SlotName", "CreatedTime", "LastPlayedTime", "Summary", "TimePlayed", "PlayCount", "LastSessionLength" }
-
-function HasSaveSlots._buildSlot(
-	self: HasSaveSlots,
-	slotId: SaveSlotData.SlotId,
-	data: SaveSlotData.SaveSlotMetadata,
-	isNew: boolean?,
-	isEphemeral: boolean?
-): ()
-	local maid = Maid.new()
-	self._maid[slotId] = maid
-
-	local slot = maid:Add(Instance.new("Folder"))
-	slot.Name = slotId
-	slot.Archivable = false
-
-	local attributes = SaveSlotData:Create(slot)
-	attributes.SlotId.Value = slotId
-	attributes.SlotIndex.Value = data.SlotIndex
-
-	if isEphemeral then
-		-- Ephemeral: seed the metadata in memory with no write-back wiring, and back the slot with an in-memory
-		-- store. The folder still joins the replicated container so the client can read the active slot's
-		-- metadata (name, summary, playtime) the same way it reads a real slot's -- what makes the slot
-		-- ephemeral is that nothing is persisted, not that it is invisible. The IsEphemeral attribute is what
-		-- keeps it out of the save list; SaveSlotDataService's list reads filter on it.
-		--
-		-- The in-memory store's lifetime is owned by this slot's maid (which is in turn owned by self._maid),
-		-- so it is Destroyed -- and becomes GC-eligible -- the instant the slot is retired (_destroyEphemeralSlot)
-		-- or the player unbinds. Nothing outside this object holds a lasting reference: while the slot is active
-		-- ObserveActiveSlotStoreBrio hands it out inside a Brio that dies when the slot deselects, releasing it.
-		attributes.IsEphemeral.Value = true
-		self._ephemeralStores[slotId] = maid:Add(InMemoryDataStore.new(slotId))
-
-		for _, key in MUTABLE_METADATA_KEYS do
-			attributes[key].Value = data[key]
-		end
-
-		-- Parent last, so the folder replicates carrying IsEphemeral -- a client that saw it land without the
-		-- flag would briefly list a throwaway slot as a save.
-		slot.Parent = self._slotContainer
-
-		self._slotMap[slotId] = slot
-		maid:GiveTask(function()
-			self._slotMap[slotId] = nil
-			self._ephemeralStores[slotId] = nil
-			self._transferableEphemeralKeys[slotId] = nil
-		end)
-		return
-	end
-
-	local metadataStore = self._metadataStore:GetSubStore(slotId)
-
-	-- Store immutable SlotIndex once on creation
-	if isNew then
-		metadataStore:Store("SlotIndex", data.SlotIndex)
-	end
-
-	-- Store mutable metadata on change
-	for _, key in MUTABLE_METADATA_KEYS do
-		attributes[key].Value = data[key]
-		maid:GiveTask(metadataStore:StoreOnValueChange(key, attributes[key]))
-
-		if isNew then
-			metadataStore:Store(key, attributes[key].Value)
-		end
-	end
-
-	slot.Parent = self._slotContainer
-
-	self._slotMap[slotId] = slot
-
-	maid:GiveTask(function()
-		self._slotMap[slotId] = nil
-	end)
-end
-
--- Whether the slot is an ephemeral (session-only, never-persisted) slot. The slot's own IsEphemeral property
--- is the single source of truth (set once at creation in _buildSlot); nil-safe so callers can pass a possibly
--- nil active-slot id directly.
-function HasSaveSlots._isEphemeral(self: HasSaveSlots, slotId: SaveSlotData.SlotId?): boolean
-	if slotId == nil then
-		return false
-	end
-	local slot = self._slotMap[slotId] :: Folder?
-	return slot ~= nil and SaveSlotData.IsEphemeral:Get(slot) == true
-end
-
--- Tears down the ephemeral slot's maid (folder, in-memory store, and every map entry via the teardown task
--- registered in _buildSlot). Idempotent -- clearing an already-cleared maid key is a no-op.
-function HasSaveSlots._destroyEphemeralSlot(self: HasSaveSlots, slotId: SaveSlotData.SlotId): ()
-	self._maid[slotId] = nil
-end
-
 function HasSaveSlots._setupSummary(self: HasSaveSlots): ()
 	self._maid:GiveTask(
 		self:ObserveActiveSlotStoreBrio():Subscribe(function(brio: Brio.Brio<DataStoreStage.DataStoreStage>)
@@ -1334,8 +570,7 @@ function HasSaveSlots._setupSummary(self: HasSaveSlots): ()
 				return
 			end
 
-			local activeSlotId = self.ActiveSlotId.Value
-			local activeSlot = activeSlotId and self._slotMap[activeSlotId]
+			local activeSlot = self._slotsDataStore:GetSlotFolder(self.ActiveSlotId.Value)
 			if not activeSlot then
 				return
 			end
@@ -1419,91 +654,8 @@ function HasSaveSlots._observeProviderValue(
 	}) :: any
 end
 
---[=[
-	Accrues per-slot playtime automatically. A "session" spans the time a slot is the active slot:
-	selecting a slot begins one (bumping PlayCount), and deselecting, switching, or unbinding ends
-	it. Elapsed wall time is folded into the slot's TimePlayed from a datastore saving callback, so
-	it persists on exactly the cadence the data is written -- always fresh at save time, with no
-	separate timer -- and again at each session boundary.
-]=]
-function HasSaveSlots._setupPlaytimeTracking(self: HasSaveSlots): ()
-	self._playSessionSlotId = nil
-	self._playSessionStart = nil
-	self._playSessionLastFlush = nil
-
-	-- The active slot bounds the session: end the previous one (if any) and begin the new one.
-	self._maid:GiveTask(self.ActiveSlotId.Changed:Connect(function()
-		self:_endPlaySession()
-
-		local activeSlotId = self.ActiveSlotId.Value
-		if activeSlotId ~= nil then
-			self:_beginPlaySession(activeSlotId)
-		end
-	end))
-
-	-- Fold accrued time into TimePlayed just before every save so the written value is current. The
-	-- callback runs before the save serializes staged data (see DataStore._syncData), so the flush is
-	-- captured by that same save -- including the final save-on-leave.
-	self._maid:GivePromise(self._loadPromise):Then(function()
-		self._maid:GiveTask(self._dataStore:AddSavingCallback(function()
-			self:_flushPlaytime()
-		end))
-	end)
-
-	-- Best-effort flush on unbind for the case where the binder tears down before that final save.
-	self._maid:GiveTask(function()
-		self:_endPlaySession()
-	end)
-end
-
-function HasSaveSlots._beginPlaySession(self: HasSaveSlots, slotId: SaveSlotData.SlotId): ()
-	local now = os.time()
-	self._playSessionSlotId = slotId
-	self._playSessionStart = now
-	self._playSessionLastFlush = now
-
-	local slot = self._slotMap[slotId]
-	if slot then
-		SaveSlotData.PlayCount:Set(slot, (SaveSlotData.PlayCount:Get(slot) or 0) + 1)
-	end
-end
-
-function HasSaveSlots._flushPlaytime(self: HasSaveSlots): ()
-	local slotId = self._playSessionSlotId
-	if slotId == nil then
-		return
-	end
-
-	local slot = self._slotMap[slotId]
-	if not slot then
-		return
-	end
-
-	local now = os.time()
-
-	-- Add only the time since the last flush so repeated flushes within a session never double count.
-	local sinceFlush = now - (self._playSessionLastFlush or now)
-	if sinceFlush > 0 then
-		SaveSlotData.TimePlayed:Set(slot, (SaveSlotData.TimePlayed:Get(slot) or 0) + sinceFlush)
-		self._playSessionLastFlush = now
-	end
-
-	SaveSlotData.LastSessionLength:Set(slot, now - (self._playSessionStart or now))
-end
-
-function HasSaveSlots._endPlaySession(self: HasSaveSlots): ()
-	if self._playSessionSlotId == nil then
-		return
-	end
-
-	self:_flushPlaytime()
-	self._playSessionSlotId = nil
-	self._playSessionStart = nil
-	self._playSessionLastFlush = nil
-end
-
 function HasSaveSlots._setupRemotes(self: HasSaveSlots): ()
-	self._maid:GiveTask(self._remoting.PromiseHasSlot:Bind(function(remotePlayer, ...)
+	self._maid:GiveTask(self._remoting.PromiseHasSlot:Bind(function(remotePlayer: Player, ...)
 		if remotePlayer == self._obj then
 			return self:PromiseHasSlot(...)
 		else
@@ -1511,7 +663,7 @@ function HasSaveSlots._setupRemotes(self: HasSaveSlots): ()
 		end
 	end))
 
-	self._maid:GiveTask(self._remoting.PromiseSelectSlot:Bind(function(remotePlayer, ...)
+	self._maid:GiveTask(self._remoting.PromiseSelectSlot:Bind(function(remotePlayer: Player, ...)
 		if remotePlayer == self._obj then
 			return self:PromiseSelectSlot(...)
 		else
@@ -1519,7 +671,7 @@ function HasSaveSlots._setupRemotes(self: HasSaveSlots): ()
 		end
 	end))
 
-	self._maid:GiveTask(self._remoting.PromiseCreateSlot:Bind(function(remotePlayer, ...)
+	self._maid:GiveTask(self._remoting.PromiseCreateSlot:Bind(function(remotePlayer: Player, ...)
 		if remotePlayer == self._obj then
 			return self:PromiseCreateSlot(...)
 		else
@@ -1527,7 +679,7 @@ function HasSaveSlots._setupRemotes(self: HasSaveSlots): ()
 		end
 	end))
 
-	self._maid:GiveTask(self._remoting.PromiseDeleteSlot:Bind(function(remotePlayer, ...)
+	self._maid:GiveTask(self._remoting.PromiseDeleteSlot:Bind(function(remotePlayer: Player, ...)
 		if remotePlayer == self._obj then
 			return self:PromiseDeleteSlot(...)
 		else
@@ -1535,7 +687,7 @@ function HasSaveSlots._setupRemotes(self: HasSaveSlots): ()
 		end
 	end))
 
-	self._maid:GiveTask(self._remoting.PromiseSetSlotMetadata:Bind(function(remotePlayer, ...)
+	self._maid:GiveTask(self._remoting.PromiseSetSlotMetadata:Bind(function(remotePlayer: Player, ...)
 		if remotePlayer == self._obj then
 			return self:PromiseSetSlotMetadata(...)
 		else
@@ -1543,7 +695,7 @@ function HasSaveSlots._setupRemotes(self: HasSaveSlots): ()
 		end
 	end))
 
-	self._maid:GiveTask(self._remoting.PromiseGetSlotMetadata:Bind(function(remotePlayer, ...)
+	self._maid:GiveTask(self._remoting.PromiseGetSlotMetadata:Bind(function(remotePlayer: Player, ...)
 		if remotePlayer == self._obj then
 			return self:PromiseGetSlotMetadata(...)
 		else
@@ -1551,7 +703,7 @@ function HasSaveSlots._setupRemotes(self: HasSaveSlots): ()
 		end
 	end))
 
-	self._maid:GiveTask(self._remoting.PromiseSlotIdFromIndex:Bind(function(remotePlayer, ...)
+	self._maid:GiveTask(self._remoting.PromiseSlotIdFromIndex:Bind(function(remotePlayer: Player, ...)
 		if remotePlayer == self._obj then
 			return self:PromiseSlotIdFromIndex(...)
 		else
@@ -1559,7 +711,7 @@ function HasSaveSlots._setupRemotes(self: HasSaveSlots): ()
 		end
 	end))
 
-	self._maid:GiveTask(self._remoting.PromiseLastActiveSlotId:Bind(function(remotePlayer)
+	self._maid:GiveTask(self._remoting.PromiseLastActiveSlotId:Bind(function(remotePlayer: Player)
 		if remotePlayer == self._obj then
 			return self:PromiseLastActiveSlotId()
 		else
